# optimize Enable: remove "enabled" metric, remove per-tag ticker, thread default tags through