--- a/flags/flags.go
+++ b/flags/flags.go
@@ -7,7 +7,7 @@ import (
 // Flag is the interface for individual feature flags
 type Flag interface {
 	FlagName() string
-	Enabled(rnd Rand, properties map[string]string) (bool, error)
+	Enabled(rnd Rand, properties, defaultTags map[string]string) (bool, error)
 	Equal(other Flag) bool
 	// Clamp returns whether a flag is always on/off, for optimization
 	Clamp() clamp.Clamp