perf is 10-72% better:

```
name                                            old time/op    new time/op    delta
Enabled/csv/50pct                                  945ns ±46%     634ns ±64%      ~     (p=0.151 n=5+5)
Enabled/csv/50pct-2                               33.7ns ± 4%    29.6ns ± 4%   -12.32%  (p=0.008 n=5+5)
Enabled/csv/50pct-4                               13.9ns ± 2%    12.7ns ± 4%    -8.44%  (p=0.008 n=5+5)
Enabled/csv/50pct-8                               7.14ns ± 2%    6.22ns ± 3%   -12.87%  (p=0.008 n=5+5)
Enabled/csv/on                                     138ns ±11%     145ns ± 6%      ~     (p=0.222 n=5+5)
Enabled/csv/on-2                                  8.31ns ± 5%    6.67ns ± 3%   -19.73%  (p=0.008 n=5+5)
Enabled/csv/on-4                                  3.73ns ± 2%    2.98ns ± 3%   -19.93%  (p=0.008 n=5+5)
Enabled/csv/on-8                                  1.91ns ± 1%    1.54ns ± 4%   -19.74%  (p=0.008 n=5+5)
Enabled/json2/50pct                                354ns ±62%     148ns ±16%      ~     (p=0.056 n=5+5)
Enabled/json2/50pct-2                             32.7ns ± 7%    28.3ns ± 5%   -13.53%  (p=0.008 n=5+5)
Enabled/json2/50pct-4                             13.5ns ± 4%    11.8ns ± 3%   -12.15%  (p=0.008 n=5+5)
Enabled/json2/50pct-8                             6.83ns ± 4%    6.04ns ± 3%   -11.58%  (p=0.008 n=5+5)
Enabled/json2/on                                  32.6ns ±16%    37.8ns ±17%      ~     (p=0.151 n=5+5)
Enabled/json2/on-2                                8.83ns ± 6%    6.76ns ± 5%   -23.37%  (p=0.008 n=5+5)
Enabled/json2/on-4                                3.63ns ± 1%    2.98ns ± 4%   -18.09%  (p=0.008 n=5+5)
Enabled/json2/on-8                                1.88ns ± 3%    1.51ns ± 2%   -19.33%  (p=0.008 n=5+5)
EnabledWithArgs/json2/flag5/map[]                  713ns ± 5%     419ns ± 7%   -41.26%  (p=0.008 n=5+5)
EnabledWithArgs/json2/flag5/map[]-2                215ns ± 5%      92ns ± 4%   -57.28%  (p=0.008 n=5+5)
EnabledWithArgs/json2/flag5/map[]-4                109ns ± 2%      43ns ± 6%   -61.00%  (p=0.008 n=5+5)
EnabledWithArgs/json2/flag5/map[]-8               76.6ns ± 4%    22.1ns ± 9%   -71.18%  (p=0.008 n=5+5)
EnabledWithArgs/json2/flag5/map[bar:b_foo:a]       865ns ± 8%     421ns ±12%   -51.33%  (p=0.008 n=5+5)
EnabledWithArgs/json2/flag5/map[bar:b_foo:a]-2     264ns ± 2%      92ns ± 2%   -64.93%  (p=0.008 n=5+5)
EnabledWithArgs/json2/flag5/map[bar:b_foo:a]-4     126ns ± 4%      42ns ± 6%   -66.65%  (p=0.008 n=5+5)
EnabledWithArgs/json2/flag5/map[bar:b_foo:a]-8    81.1ns ± 4%    22.0ns ±10%   -72.92%  (p=0.008 n=5+5)

name                                            old alloc/op   new alloc/op   delta
Enabled/csv/50pct                                   268B ±50%      176B ±69%      ~     (p=0.151 n=5+5)
Enabled/csv/50pct-2                                3.40B ±18%     3.00B ± 0%      ~     (p=0.333 n=5+4)
Enabled/csv/50pct-4                                1.00B ± 0%     1.00B ± 0%      ~     (all equal)
Enabled/csv/50pct-8                                0.00B          0.00B           ~     (all equal)
Enabled/csv/on                                     36.8B ±11%     40.4B ± 6%      ~     (p=0.056 n=5+5)
Enabled/csv/on-2                                  0.60B ±100%     0.00B           ~     (p=0.167 n=5+5)
Enabled/csv/on-4                                   0.00B          0.00B           ~     (all equal)
Enabled/csv/on-8                                   0.00B          0.00B           ~     (all equal)
Enabled/json2/50pct                                56.2B ±71%     20.2B ±21%      ~     (p=0.222 n=5+5)
Enabled/json2/50pct-2                              2.00B ± 0%     1.60B ±38%      ~     (p=0.333 n=4+5)
Enabled/json2/50pct-4                             0.40B ±150%     0.00B           ~     (p=0.444 n=5+5)
Enabled/json2/50pct-8                              0.00B          0.00B           ~     (all equal)
Enabled/json2/on                                   3.40B ±18%     4.80B ±25%      ~     (p=0.079 n=5+5)
Enabled/json2/on-2                                 0.00B          0.00B           ~     (all equal)
Enabled/json2/on-4                                 0.00B          0.00B           ~     (all equal)
Enabled/json2/on-8                                 0.00B          0.00B           ~     (all equal)
EnabledWithArgs/json2/flag5/map[]                   436B ± 2%       53B ±13%   -87.95%  (p=0.008 n=5+5)
EnabledWithArgs/json2/flag5/map[]-2                 377B ± 1%        7B ±18%   -98.20%  (p=0.008 n=5+5)
EnabledWithArgs/json2/flag5/map[]-4                 366B ± 0%        2B ± 0%   -99.45%  (p=0.016 n=5+4)
EnabledWithArgs/json2/flag5/map[]-8                 362B ± 0%        1B ± 0%      ~     (p=0.079 n=4+5)
EnabledWithArgs/json2/flag5/map[bar:b_foo:a]        452B ± 3%       53B ±17%   -88.26%  (p=0.008 n=5+5)
EnabledWithArgs/json2/flag5/map[bar:b_foo:a]-2      383B ± 1%        7B ± 0%   -98.17%  (p=0.000 n=5+4)
EnabledWithArgs/json2/flag5/map[bar:b_foo:a]-4      367B ± 1%        2B ± 0%   -99.46%  (p=0.016 n=5+4)
EnabledWithArgs/json2/flag5/map[bar:b_foo:a]-8      363B ± 0%        1B ± 0%   -99.72%  (p=0.008 n=5+5)
```

(this compares master + the first commit of this series, which is a benchmark-only change, to the tip of the branch)

It is probably best to review this PR by PR, but at a high level:
* we stop emitting the "enabled" tag, as users who care are passing in an enabledCB func.
* that enables us to remove creating and managing `time.Ticker`s for each flag
* that enables us to simplfy the logic in Enable, where we can move from checking if there is something to read from the channel (which requires taking a mutex) to checking a boolean, which should scale better at high core counts.
* we can avoid allocating in hashValue by allocating a buffer for the sha1 sum that the compiler identifies as not escaping, and ends up allocating on the stack
* finally (and TBH probably most impactful) we can thread through `defaultTags` to the flags2 backend rather than merging properties and defaultTags into 1 map, avoiding a big source of allocations.